disable vagrant ssh key replacement
Vagrant replaces the insecure default private key, but doesn't
understand that boot2docker uses tmpfs, so the private key reverts on
next reboot, and vagrant can no longer ssh in.

This isn't an ideal fix, I'd like to support a custom private key
eventually, but it's equivalent to vagrant 1.6 behavior.