squash NFS permissions to the home dir user
This is less compatible with running Docker directly on Linux, where no
squashing occurs on host-mounted volumes, but it's a lot more convenient for
users and the edge cases where it makes a difference appear to be quite rare.

Example of an edge case where this breaks:

    > mkdir pg1 && docker run -v `pwd`/pg1:/var/lib/postgresql/data postgres
    chown: changing ownership of `/var/lib/postgresql/data': Operation not permitted

This fails because postgres requires the data dir to be owned by the container's postgres user.

This doesn't work on boot2docker either, with virtualbox built-in file sharing,
but for a different reason -- on dinghy with NFS, it fails immediately because
the `chown` isn't allowed. On boot2docker with virtualbox sharing, the `chown`
is silently ignored, and then you get an error later because the directory isn't
owned by the container's "postgres" user.