Fix FS Events forwarding to xhyve VM
Just noticed that today and I tracked down the root cause to a blanky
`SSHKeyPath` on the VM's state that results in wrong arguments being
passed on to `fsevents_to_vm`:

```
$ docker-machine create remove --driver xhyve
....
Docker is up and running!
To see how to connect your Docker Client to the Docker Engine running on
this virtual machine, run: docker-machine env remove

$ docker-machine inspect remove | grep -i sshkey
        "SSHKeyPath": "",
```

This might be something to be fixed on the driver's side but I guess it
does the trick for now.