avoid code that may break in the sudo environment
This fixes issues with the sudo environment being configured differently
on different machines, especially with the whitelisted env vars to keep
in the env.

This involved a refactor of the dinghy constants, primarily to increase
confidence that we aren't accidentally relying on state in the sudo'd
process, but also for general code cleanup.