work around DNS resolution failing for a few seconds
After configuring the new resolver, we're seeing a fairly common problem
with DNS resolution failing in the VM for a second or two. This seems to
be new with MacOS 10.12, but we can't be certain.

It might be nice to actually poll DNS over SSH in a loop until it starts
working, but for now we're just throwing in a silly sleep(5).