don't load ~/.gemrc when installing fsevents_to_vm
This fix avoid loading the current user's gemrc both when probing if
`fsevents_to_vm` installed, and when installing it. If the gem is installed
using `--env-shebang` an incorrect ruby may be located (by searching the user's
`PATH`) resulting in a failure to start the FsEvents daemon.

Related to issue #227. `~/.gemrc` contained:

```
install: --env-shebang
```